moved events to utils folder
