+++ frontend/src/components/layout/grids/WeaponGrid.vue
@@ -191,7 +191,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import Events from '../../../events';
+import Events from '../../../utils/events';
 import gasp from 'gsap';
 import { Accessors, PropType } from 'vue/types/options';
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex';
