+++ frontend/src/components/stake/StakeSelectorItem.vue
@@ -154,7 +154,7 @@ import BN from 'bignumber.js';
 import {mapActions, mapState} from 'vuex';
 import i18n from '@/i18n';
 import { TranslateResult } from 'vue-i18n';
-import Events from '../../events';
+import Events from '../../utils/events';
 import SkeletonLoader from '../loader/SkeletonLoader.vue';
 import { stakeTypeThatCanHaveUnclaimedRewardsStakedTo } from '@/stake-types';
