+++ frontend/src/views/Options.vue
@@ -101,7 +101,7 @@
 </template>
 
 <script lang="ts">
-import Events from '../events';
+import Events from '../utils/events';
 import {mapActions, mapGetters, mapMutations, mapState} from 'vuex';
 import BigNumber from 'bignumber.js';
 import {Accessors} from 'vue/types/options';
