+++ frontend/src/views/SelectStakeType.vue
@@ -26,7 +26,7 @@
 import { mapActions, mapGetters, mapState } from 'vuex';
 import BN from 'bignumber.js';
 import _ from 'lodash';
-import Events from '../events';
+import Events from '../utils/events';
 BN.config({ ROUNDING_MODE: BN.ROUND_DOWN });
 BN.config({ EXPONENTIAL_AT: 100 });
 import StakeSelectorItem from '../components/stake/StakeSelectorItem.vue';
